requirements.txt
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.1
pytz==2024.1
SQLAlchemy==2.0.25
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
Flask-WTF==1.2.1 